% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{parse_silo_csv}
\alias{parse_silo_csv}
\title{Parse a SILO Data Drill CSV}
\usage{
parse_silo_csv(path)
}
\arguments{
\item{path}{Path to a SILO CSV response.}
}
\value{
Data frame with standard RCLT climate fields.
}
\description{
Parse a SILO Data Drill CSV
}
