% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\name{parse_slga_bulk_density}
\alias{parse_slga_bulk_density}
\title{Parse normalized SLGA whole-earth bulk density}
\usage{
parse_slga_bulk_density(path)
}
\arguments{
\item{path}{Path to a normalized SLGA JSON cache.}
}
\value{
Data frame containing depth bands and density estimates in g/cm3.
}
\description{
Parse normalized SLGA whole-earth bulk density
}
