% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_climate_forcing}
\alias{plot_climate_forcing}
\title{Plot daily climate and infiltration forcing}
\usage{
plot_climate_forcing(forcing)
}
\arguments{
\item{forcing}{Data frame containing rain, PET, daily infiltration, cumulative
infiltration, and either dates or elapsed time.}
}
\value{
The input data, invisibly.
}
\description{
Plot daily climate and infiltration forcing
}
