% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mass_fractions}
\alias{plot_mass_fractions}
\title{Plot CLTF mass fractions through time}
\usage{
plot_mass_fractions(simulation)
}
\arguments{
\item{simulation}{Simulation table returned by \code{\link[=simulate_cltf]{simulate_cltf()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot CLTF mass fractions through time
}
