% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_objective_profile}
\alias{plot_objective_profile}
\title{Plot one or more objective profiles}
\usage{
plot_objective_profile(profile)
}
\arguments{
\item{profile}{Profile table returned by \code{\link[=profile_rclt_parameter]{profile_rclt_parameter()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot one or more objective profiles
}
