% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot log residuals against fitted concentration}
\usage{
plot_residuals(predictions)
}
\arguments{
\item{predictions}{Fit prediction table returned by \code{\link[=fit_rclt]{fit_rclt()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot log residuals against fitted concentration
}
