% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{round_silo_coordinate}
\alias{round_silo_coordinate}
\title{Round a coordinate to the SILO 0.05-degree grid}
\usage{
round_silo_coordinate(value)
}
\arguments{
\item{value}{Decimal-degree coordinate.}
}
\value{
Rounded coordinate.
}
\description{
Round a coordinate to the SILO 0.05-degree grid
}
