% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mass_fractions}
\alias{plot_mass_fractions}
\title{Plot RCLT mass fractions through time}
\usage{
plot_mass_fractions(simulation)
}
\arguments{
\item{simulation}{Simulation table returned by \code{\link[=simulate_rclt]{simulate_rclt()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot RCLT mass fractions through time
}
